Utils reorganization
