+++ project/src/components/home-page-list/home-page-list.tsx
@@ -2,13 +2,13 @@ import React, { useCallback, useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { CardArticleClasses, CardImgWrapperClasses, FetchStatus, SortOptions } from '../../const';
 import { getActiveCity, getFetchStatus } from '../../store/selectors';
-import { getSortedData } from '../../utils/common-utils';
 import FetchFailMessage from '../fetch-fail-message/fetch-fail-message';
 import HomePageSortDropdown from '../home-page-sort-dropdown/home-page-sort-dropdown';
 import HomePageSortToggler from '../home-page-sort-toggler/home-page-sort-toggler';
 import { OfferDataTypes } from '../../types/offer-data-types';
 import OfferCard from '../offer-card/offer-card';
 import SpinnerHome from '../spinner-home/spinner-home';
+import { getSortedData } from '../../utils/project-specific-utils';
 
 type HomePageListTypes = {
   offersData: OfferDataTypes[],
