+++ project/src/components/offer-page-main/offer-page-main.tsx
@@ -4,7 +4,6 @@ import { useParams } from 'react-router-dom';
 import { MAX_IMAGES_NUMBER } from '../../const';
 import { getCurrentHotel, getNearOffersData } from '../../store/selectors';
 import { getNearOffersAction, getOfferDataAction } from '../../store/api-actions';
-import { getVisualRating } from '../../utils/common-utils';
 import OfferPageGallery from '../offer-page-gallery/offer-page-gallery';
 import OfferPageGoods from '../offer-page-goods/offer-page-goods';
 import OfferPageHost from '../offer-page-host/offer-page-host';
@@ -12,6 +11,7 @@ import OfferPageMap from '../offer-page-map/offer-page-map';
 import OfferPageNearList from '../offer-page-near-list/offer-page-near-list';
 import OfferPageBookmark from '../offer-page-bookmark/offer-page-bookmark';
 import OfferPageReviews from '../offer-page-reviews/offer-page-reviews';
+import { getVisualRating } from '../../utils/project-specific-utils';
 
 function OfferPageMain(): JSX.Element {
   const pageData = useSelector(getCurrentHotel);
